Add blank lines between headings and lists in release notes
Fixes the five MD032 (blanks-around-lists) failures in the markdownlint job.
These pre-date this branch; CI surfaces them because generate-lockfile
re-resolves pymarkdownlnt ("~0.9.30") to 0.9.37, which detects the pattern
that 0.9.36 missed.
