+++ docs/admin/release_notes/version_0.7.md
@@ -17,6 +17,7 @@
 ### Added
 
 ### Changed
+
 - [#73](https://github.com/nautobot/nautobot-app-bgp-models/pull/73) - Documentation Update
 - [#90](https://github.com/nautobot/nautobot-app-bgp-models/pull/90) - Development environments updates
